pyproject.toml
[tool.poetry]
name = "simple-web-service"
version = "0.1.0"
description = ""
authors = ["Aziza <azizamukashkyzy@gmail.com>"]
readme = "README.md"
packages = [{include = "simple_web_service"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
djangorestframework = "^3.14.0"
pytest = "^7.2.2"
factory-boy = "^3.2.1"
faker = "^17.6.0"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
